% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph.R
\name{mph_as_factor}
\alias{mph_as_factor}
\title{Convert string vector to a factor}
\usage{
mph_as_factor(s)
}
\arguments{
\item{s}{strings}
}
\value{
factor
}
\description{
Factor levels are in "first seen" order i.e. not sorted alphabetically.
}
\examples{
mph_as_factor(letters)
mph_as_factor(c('h', 'e', 'l', 'l', 'o'))
}
