% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph.R
\name{mph_init}
\alias{mph_init}
\title{Initialise a hash object with the given set of strings}
\usage{
mph_init(s, size_factor = 1, verbosity = 0L)
}
\arguments{
\item{s}{character vector. Duplicates and NAs are not 
allowed.}

\item{size_factor}{Adjusts the internal number of hash slots. Default: 1. Increasing 
this factor lowers the chance of hash collisions and may speed up the
overall performance.}

\item{verbosity}{What level of debuggin information should be output? Default: 0
(no debugging output). Valid levels: 0, 1, 2}
}
\value{
Hash object
}
\description{
Initialise a hash object with the given set of strings
}
\examples{
mph <- mph_init(letters)
mph_match(c('h', 'e', 'l', 'l', 'o'), mph)
}
