% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph.R
\name{mph_match}
\alias{mph_match}
\title{Find the integer index of string elements.}
\usage{
mph_match(s, mph)
}
\arguments{
\item{s}{strings}

\item{mph}{hash object created by \code{mph_init()}}
}
\value{
integer vector of indexes for each of the strings. 
   NA if string is not a member.
}
\description{
This is equivalent to R's \code{match()} function.
}
\examples{
mph <- mph_init(letters)
mph_match(c('h', 'e', 'l', 'l', 'o'), mph)
}
