% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph.R
\name{mph_subset}
\alias{mph_subset}
\title{Subset an R object using a hashed index}
\usage{
mph_subset(nms, x, mph)
}
\arguments{
\item{nms}{names of elements to extract}

\item{x}{Object to subset. Either list or vector (integer, numeric or character)}

\item{mph}{hash object created by \code{mph_init()}}
}
\value{
Subset of elements of x
}
\description{
Subset an R object using a hashed index
}
\examples{
mph <- mph_init(letters)
values <- LETTERS
mph_match(c('h', 'e', 'l', 'l', 'o'), mph)
mph_subset(c('h', 'e', 'l', 'l', 'o'), values, mph)
}
