% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph-interface.R
\name{mph_init}
\alias{mph_init}
\title{Initialise a minimal perfect hash with the given set of strings}
\usage{
mph_init(s)
}
\arguments{
\item{s}{character vector of all possible strings. Duplicates and NAs not 
allowed.  String length must be less than 255 characters.}
}
\value{
mph object
}
\description{
Initialise a minimal perfect hash with the given set of strings
}
\examples{
mph <- mph_init(letters)
mph_match(c('h', 'e', 'l', 'l', 'o'), mph)
}
