% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph.R
\name{mph_match}
\alias{mph_match}
\title{Find the integer index of string elements}
\usage{
mph_match(s, mph)
}
\arguments{
\item{s}{strings to find the index for}

\item{mph}{minimal perfect hash object created by \code{mph_init()}}
}
\value{
integer vector of indicies of strings within the mph object. 
   NA if string is not a member of \code{mph} object.
}
\description{
Find the integer index of string elements
}
\examples{
mph <- mph_init(letters)
mph_match(c('h', 'e', 'l', 'l', 'o'), mph)
}
