% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph-interface.R
\name{mph_subset}
\alias{mph_subset}
\title{Subset an integer vector, numeric vector or list using a hashed index}
\usage{
mph_subset(x, s, mph)
}
\arguments{
\item{x}{integer vector, numeric vector or list}

\item{s}{strings to search for}

\item{mph}{minimal perfect hash of string population}
}
\value{
Subset of elements of x
}
\description{
Subset an integer vector, numeric vector or list using a hashed index
}
\examples{
mph <- mph_init(letters)
value <- 2^seq_along(letters)
mph_match(c('h', 'e', 'l', 'l', 'o'), mph)
mph_subset(value, c('h', 'e', 'l', 'l', 'o'), mph)
}
