% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph.R
\name{mph_subset}
\alias{mph_subset}
\title{Subset an integer vector, numeric vector or list using a hashed index}
\usage{
mph_subset(elems, x, mph)
}
\arguments{
\item{elems}{names of elements to extract}

\item{x}{Object to be subset. One of: integer vector, numeric vector or list}

\item{mph}{hash obbject}
}
\value{
Subset of elements of x
}
\description{
Subset an integer vector, numeric vector or list using a hashed index
}
\examples{
mph <- mph_init(letters)
values <- 2^seq_along(letters)
mph_match(c('h', 'e', 'l', 'l', 'o'), mph)
mph_subset(c('h', 'e', 'l', 'l', 'o'), values, mph)
}
